+++ app/ospf/components/core/MenuLayout/MenuLayout.js
@@ -1,10 +1,10 @@
 class MenuLayout extends Component {
     async init({}={}) {
         await super.init();
-        this.contents = [];
+        this.tabs = [];
 
-        let contentPlaceholder = await this.addChild("content", new ComponentPlaceholder());
-        await contentPlaceholder.init({
+        let contentWrapper = await this.addChild("content", new ComponentWrapper());
+        await contentWrapper.init({
             className: "TabbedLayout_content TabbedLayout_content-menu"
         });
 
@@ -15,8 +15,8 @@ class MenuLayout extends Component {
     async onEvent(event){
         if(event.name == "TabsBar_tabClick"){
             if(event.sender == this.getChild("tabsBar")){
-                let component = this.contents.find(item => item.key == event.parameter.key);
-                await this.setContent(component.component);
+                let tab = this.tabs.find(item => item.key == event.parameter.key);
+                await this.setContent(tab.component);
             }
         }
     }
@@ -32,7 +32,7 @@ class MenuLayout extends Component {
             key: key
         });
 
-        this.contents.push({
+        this.tabs.push({
             component: component,
             key: key
         });